Big changes
Removed all the various Skin variants. Added a single BatchInfo to Skins
in order to minimize the number of BatchInfos around. Maybe it will be
better to assign a Material directly and use that one instead of
generating the BatchInfo by myself.

Still some issues on plugin reload, but greatly improved in the
management of newly made Widgets.